Disable /suicide and /kill commands for players in minigame rounds
These commands can cause issues because they don't always allow the death event to be overridden, resulting
in very strange and destructive behavior. Disabling them is the easiest solution as of yet. /kill will still
work provided an argument which is not a player in a round is given. /suicide will be blocked entirely.