Only load deprecated block entity classes if on a legacy MC version
This avoids NoClassDefFoundErrors when attempting to serialize or
deserialize blocks on newer Bukkit versions which no longer provide
these deprecated classes. The (de)serialization code was only running
on legacy versions anyway - all this commit does is refactor the code
to first check the Minecraft version so the instanceof check can be
skipped if it's unnecessary.